feat: 동기적으로 auth 상태를 참조하여, stale한 인증상태들을 동기적으로 참조하도록 변경
