+++ src/bounded-context/user/components/loginUserMenu/UserWidget.tsx
@@ -12,12 +12,16 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
+import { useSession } from "@/global/hook/useSession";
 
 
 
 
 
 export default function UserWidget({className} : {className?: string}){
+
+  const { session, logout } = useSession();
+
   return (
     <div className={className}>
       <DropdownMenu modal>
@@ -45,7 +49,7 @@ export default function UserWidget({className} : {className?: string}){
           <DropdownMenuSeparator />
           <DDLink href="/account" name="계정 설정" icon={<UserCog />} />
           <DDLink href="/support" name="고객센터" icon={<Phone />} />
-          <DDLink href="/logout" name="로그아웃" icon={<LogOut />} />
+          <DDAction onClick={logout} name="로그아웃" icon={<LogOut />} />
         </DropdownMenuContent>
       </DropdownMenu>
     </div>
@@ -54,11 +58,19 @@ export default function UserWidget({className} : {className?: string}){
 
 // Dropdown Menu의 item중, Link 컴포넌트 기반 + 아이콘인 경우
 function DDLink({href, name, icon} : {href: string, name: string, icon: React.ReactNode}){
-  return (
+  return(
     <Link href={href} className="flex items-center">
       <DropdownMenuItem className="w-full cursor-pointer">
         {icon}<span className="ml-2">{name}</span>
       </DropdownMenuItem>
     </Link>
+  )
+}
+
+function DDAction({onClick, name, icon} : {onClick: () => void, name: string, icon: React.ReactNode}){
+  return(
+    <DropdownMenuItem className="w-full cursor-pointer" onClick={onClick}>
+        {icon}<span className="ml-2">{name}</span>
+    </DropdownMenuItem>
   )
 }
