+++ src/global/config/next-dynamic-config.ts
@@ -1,6 +1,6 @@
 import dynamic from 'next/dynamic';
 
 export const RootAuthFromLocalStorage = dynamic(
-  () => import('@/global/hook/useRootAuth').then(mod => mod.useRootAuth), 
+  () => import('@/global/hook/useAuth').then(mod => mod.useAuth), 
   { ssr: false }
 );
