+++ src/global/hook/useAuth.ts
@@ -1,115 +1,68 @@
-import { AccessToken, AccessTokenStorage, PrivatePropertyAccessTokenStorage } from "@/bounded-context/user/service/AccessTokenStorage";
+'use client';
 import { create } from "zustand";
-import { DefaultAxiosFactory } from "@/global/api/DefaultAxiosFactory";
-import { triggerToast } from "@/global/toast/ToastProvider";
-import { useRootAuth } from "./useRootAuth";
-import { AsyncRequestBuilder } from "@/global/api/AsyncRequestBuilder";
+import { triggerToast } from "../toast/ToastProvider";
 
-/* AccessTokenStorage을 전역 싱글턴 인스턴스로 관리하기 위한 zustand store
- * 다만, 접근은 useAccessToken() 훅을 통해서만 한다. (export X)
- */
-interface AccessTokenStore {
-  storage: AccessTokenStorage
-}
-const useAccessTokenStore = create<AccessTokenStore>((set, get) => ({
-  storage: new PrivatePropertyAccessTokenStorage()
-}));
+export const SESSION_FLAG_NAME = "_pageflowSessionExpiredAt" as const;
 
-export const useAuth = () => {
-  const { isAuthenticated } = useRootAuth(); // 인증 상태 참조
-  const { storage } = useAccessTokenStore(); // 저장소 참조
+function isAuthenticated(){
+  const isAuthenticated = continueIfOnServer(() => {
+    const sessionFlag = getLocalStorage().getItem(SESSION_FLAG_NAME);
 
-  const storeToken = (token: AccessToken) => {
-    useAccessTokenStore.getState().storage.store(token);
-  }
-
-  const clearToken = () => {
-    useAccessTokenStore.getState().storage.clear();
-  }
-
-
-  const ensureToken: () => Promise<string> = async () => {
+    // [1]: 세션지표가 존재하지 않는 경우 -> refreshToken 쿠키는 존재할 수도 있지만, 모르겠고 그냥 세션 없는걸로 간주
+    if(!sessionFlag) {
+      return false;
+    }
     
-    const isTokenExist: boolean = storage.isTokenExist();
-    // [1]: 요청결정 변수: 해당 값을 최종확인하여, refresh 요청을 전송할지를 결정.
-    let isRefreshRequired: boolean = false;
-  
-  
-    // [2]: 토큰 상태 검증
-    // 토큰 있음 => 만료여부 확인 
-    if(isTokenExist){
-      if(storage.isTokenExpired()){
-        isRefreshRequired = true;
-      }
-    /**
-     * 토큰 없음 -> isAuthentication이 true라면, refresh 요청을 보냄
-     */
-    } else {
-      if(isAuthenticated){
-        isRefreshRequired = true;
-      }
+    // [2]: 세션지표가 존재하지만, 만료된 경우 -> 세션 없음
+    const isSessionExpired = Number(sessionFlag) < Date.now();
+    if(isSessionExpired) {
+      return false;
     }
-  
-    // [3]: accessToken이 없다면 refresh 요청을 보내고, 새로운 토큰을 받아옴
-    if(isRefreshRequired){
-      const request = new AsyncRequestBuilder(DefaultAxiosFactory.createDefaultAxiosInstance(), null);
-      console.debug("refresh 요청 전송됨...");
-      
-      const accessToken: AccessToken = await request.post("/refresh")
-        .actions({
-          TOKEN_NOT_FOUND, // refreshTokenUUID 쿠키가 없는 경우 -> 세션 없음
-          SESSION_EXPIRED // refreshTokenUUID 쿠키가 있지만, 만료된 경우 -> 세션 만료
-        })
-        .fetch<AccessToken>();
-  
-      // 받아온 토큰 저장 후 반환.
-      storage.store(accessToken);
-      return accessToken.compact;
-  
-      // [4]: refresh 요청을 보내지 않고, 유효한 기존의 토큰을 그대로 반환
-      } else {
-        return storage.getToken();
-      }
-  }
+    
+    // [3]: 세션지표가 존재하고, 만료되지 않은 경우 -> 세션 있음
+    return true;
+  })
+  console.debug("LocalStorage에서 ROOT 인증상태를 불러옴: ", isAuthenticated);
+  return isAuthenticated;
+}
 
-  return {
-    storeToken, 
-    clearToken,
-    ensureToken
+function authenticate(){
+  continueIfOnServer(() => {
+    const expiredAt = Date.now() + 1000 * 60 * 60 * 24 * 30; // 30일
+    localStorage.setItem(SESSION_FLAG_NAME, String(expiredAt));
+  })
+}
+
+function deAuthenticate(){
+  continueIfOnServer(() => localStorage.removeItem(SESSION_FLAG_NAME))
+}
+
+function getLocalStorage(): Storage {
+  return window.localStorage;
+}
+
+function continueIfOnServer(callback: any){
+  while(typeof window !== "undefined" && typeof window.localStorage !== "undefined"){
+    return callback();
   }
 }
 
-
-
-
-
-
-
-
-// CodeActions
-function TOKEN_NOT_FOUND(){
-  // 로그아웃 처리
-  useRootAuth().logout();
-
-  triggerToast({
-    title: "로그인이 필요합니다",
-    action: {
-      description: "로그인하러 가기",
-      onClick: () => { console.log("로그인 페이지로 이동"); }
-    }
-  });
+export interface RootAuth {
+  isAuthenticated: boolean;
+  authenticate: () => void;
+  deAuthenticate: () => void;
 }
 
-function SESSION_EXPIRED(){
-  // 로그아웃 처리
-  useRootAuth().logout();
+export const useAuth = create<RootAuth>((set, get) => ({
+  isAuthenticated: isAuthenticated(),
 
-  triggerToast({
-    title: "만료된 세션입니다. 다시 로그인 해주세요.",
-    action: {
-      description: "로그인하러 가기",
-      onClick: () => { console.log("로그인 페이지로 이동"); }
-    }
-  });
-}
+  authenticate: () => {
+    authenticate();
+    set({ isAuthenticated: true });
+  },
 
+  deAuthenticate: () => {
+    deAuthenticate();
+    set({ isAuthenticated: false });
+  }
+}));
